requirements.txt
python-dotenv
pipecat-ai[cartesia,groq,silero,websocket,deepgram,elevenlabs]